Rounding Error Fixed, integer math converted to double - changed commit user

+++ FinalExamGradeCalculator/src/gradecalculator/model/Assignment.java
@@ -1,9 +1,9 @@
 package gradecalculator.model;
 
 public class Assignment {
-	int pctgrade;
-	int pcttotalworth;
-	public Assignment(int pctgrade, int pcttotalworth){
+	double pctgrade;
+	double pcttotalworth;
+	public Assignment(double pctgrade, double pcttotalworth){
 		this.pctgrade = pctgrade;
 		this.pcttotalworth = pcttotalworth;
 	}
@@ -11,14 +11,14 @@ public class Assignment {
 	 * Getter for Percent Grade.
 	 * @return
 	 */
-	public int getPctGrade(){
+	public double getPctGrade(){
 		return pctgrade;
 	}
 	/**
 	 * Getter for Percent of the Total Worth.
 	 * @return
 	 */
-	public int getPctTotalWorth(){
+	public double getPctTotalWorth(){
 		return pcttotalworth;
 	}
 }
